@@ -1,6 +1,6 @@
 {
   "name": "svelteup",
-  "version": "3.0.0",
+  "version": "4.0.0",
   "description": "A command line is used to bundle svelte components as web component default.",
   "module": "dist/index.js",
   "main": "dist/index.cjs",
@@ -9,8 +9,6 @@
     "prepare": "tsup src/index.ts --format esm,cjs",
     "build": "tsup src/index.ts --format esm,cjs --dts",
     "commit": "git-cz",
-    "example1": "svelteup -c examples/default/svelteup.config.js",
-    "example2": "svelteup -c examples/no-custom-element/svelteup.config.ts",
     "test": "uvu tests test.mjs -i setup",
     "format": "prettier --write --plugin-search-dir=. ."
   },
@@ -35,36 +33,41 @@
   "homepage": "https://github.com/brandonxiang/svelteup#readme",
   "dependencies": {
     "sade": "^1.8.1",
-    "esbuild": "0.19.2",
-    "esbuild-svelte": "0.8.0",
-    "bundle-require": "^4.0.1",
-    "fast-glob": "^3.2.11"
+    "esbuild": "0.24.0",
+    "esbuild-svelte": "0.9.0",
+    "bundle-require": "^5.0.0",
+    "fast-glob": "^3.3.2"
   },
   "devDependencies": {
-    "svelte": "^4.2.1",
-    "@types/sade": "^1.7.4",
-    "@types/node": "^18.6.2",
+    "svelte": "^5.1.16",
+    "@types/sade": "^1.8.0",
+    "@types/node": "^22.9.0",
     "git-cz": "^4.9.0",
-    "tsup": "^7.2.0",
-    "typescript": "^5.2.2",
+    "tsup": "^8.3.5",
+    "typescript": "^5.6.3",
     "@typescript-eslint/parser": "^6.7.2",
     "@typescript-eslint/eslint-plugin": "^6.7.2",
     "uvu": "^0.5.6",
     "puppeteer": "^15.5.0",
-    "svelte-preprocess": "^5.0.4",
-    "postcss": "^8.4.14",
-    "autoprefixer": "^10.4.8",
+    "svelte-preprocess": "^6.0.3",
+    "postcss": "^8.4.49",
+    "autoprefixer": "^10.4.20",
     "eslint": "^8.50.0",
-    "eslint-plugin-svelte": "^2.33.2",
-    "prettier-plugin-svelte": "^3.0.3",
-    "prettier": "^3.0.3",
-    "sirv": "^2.0.3",
-    "release-it": "^16.2.0"
+    "eslint-plugin-svelte": "^2.45.1",
+    "prettier-plugin-svelte": "^3.2.6",
+    "prettier": "^3.1.0",
+    "sirv": "^3.0.0",
+    "release-it": "^17.10.0"
   },
   "bin": {
     "svelteup": "bin.js"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org/"
+  },
+  "pnpm": {
+    "overrides": {
+      "rollup": "npm:@rollup/wasm-node"
+    }
   }
 }
