@@ -1,27 +1,21 @@
 <svelte:options customElement="crud-app" />
 
 <script>
-  let people = [
+  import { run } from 'svelte/legacy';
+
+  let people = $state([
     { first: 'Hans', last: 'Emil' },
     { first: 'Max', last: 'Mustermann' },
     { first: 'Roman', last: 'Tisch' },
-  ];
+  ]);
 
-  let prefix = '';
-  let first = '';
-  let last = '';
-  let i = 0;
+  let prefix = $state('');
+  let first = $state('');
+  let last = $state('');
+  let i = $state(0);
 
-  $: filteredPeople = prefix
-    ? people.filter((person) => {
-        const name = `${person.last}, ${person.first}`;
-        return name.toLowerCase().startsWith(prefix.toLowerCase());
-      })
-    : people;
 
-  $: selected = filteredPeople[i];
 
-  $: reset_inputs(selected);
 
   function create() {
     people = people.concat({ first, last });
@@ -48,6 +42,16 @@
     first = person ? person.first : '';
     last = person ? person.last : '';
   }
+  let filteredPeople = $derived(prefix
+    ? people.filter((person) => {
+        const name = `${person.last}, ${person.first}`;
+        return name.toLowerCase().startsWith(prefix.toLowerCase());
+      })
+    : people);
+  let selected = $derived(filteredPeople[i]);
+  run(() => {
+    reset_inputs(selected);
+  });
 </script>
 
 <input placeholder="filter prefix" bind:value={prefix} />
@@ -62,11 +66,11 @@
 <label><input bind:value={last} placeholder="last" /></label>
 
 <div class="buttons">
-  <button on:click={create} disabled={!first || !last}>create</button>
-  <button on:click={update} disabled={!first || !last || !selected}
+  <button onclick={create} disabled={!first || !last}>create</button>
+  <button onclick={update} disabled={!first || !last || !selected}
     >update</button
   >
-  <button on:click={remove} disabled={!selected}>delete</button>
+  <button onclick={remove} disabled={!selected}>delete</button>
 </div>
 
 <style>
