@@ -25,6 +25,6 @@ sade('svelteup [entry]', true)
     '[Development] Watch Mode without serving static resources (default false)',
   )
   .option('-s, --servedir', '[Development] Static resources directory')
-  .option('-p, --port', '[Development] Serve port (default 5000)')
+  .option('-p, --port', '[Development] Serve port (default 9527)')
   .action(svelteup)
   .parse(process.argv, {});
