@@ -1,7 +1,7 @@
 <svelte:options customElement="counter-app" />
 
 <script>
-  let count = 1;
+  let count = $state(1);
 
   function onclick() {
     count += 1;
@@ -9,4 +9,4 @@
 </script>
 
 <input type="number" bind:value={count} />
-<button on:click={onclick}>count </button>
+<button {onclick}>count </button>
